\name{plotProjection}
\alias{plotProjection}
\title{Overview of the Sample Expression}
\description{
Plots a projection of sample expression profiles on the differential expression signature.
}
\usage{
plotProjection(projection, label, col, colourMap)
}
\arguments{
  \item{projection}{a vector of scalar products between each sample's expression profile and differential expression signature}
  \item{label}{character vector of sample names}
  \item{col}{sample colour palette}
  \item{colourMap}{sample colour map}
}
\value{
The result of the function is a plot.
}
\author{Wolfgang Huber, 2013}
\examples{
## For illustration of use please see the package vignette
}
